# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Instrumentos Musicais",
    "version": "1.0.0",
    "description": "API para cadastro e visualização de istrumentos musicais"
  },
  "paths": {
    "/instrumentos": {
      "get": {
        "summary": "Obter a lista de instrumentos",
        "responses": {
          "200": {
            "description": "Lista de instrumentos retornada com sucesso"
          }
        }
      },
      "post": {
        "summary": "Criar um novo instrumento",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Instrumentos"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Instrumento criado com sucesso"
          },
          "400": {
            "description": "Dados incompletos fornecidos"
          }
        }
      }
    },
    "/instrumentos/{id}": {
      "put": {
        "summary": "Atualizar informações de um instrumento pelo ID",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Instrumentos"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Informações do instrumento atualizadas com sucesso"
          },
          "404": {
            "description": "Instrumento não encontrado"
          }
        }
      },
      "delete": {
        "summary": "Excluir um instrumento pelo ID",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Instrumento excluído com sucesso"
          },
          "404": {
            "description": "Instrumento não encontrado"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Instrumentos": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer"
          },
          "nome": {
            "type": "string"
          },
          "imagem": {
            "type": "string",
            "format": "url"
          },
          "familia": {
            "type": "string"
          },
          "tipo": {
            "type": "string"
          },
          "material": {
            "type": "string"
          },
          "origem": {
            "type": "string"
          },
          "descricao": {
            "type": "string"
          }
        }
      }
    }
  }
}